# public/swagger/swagger.yml
openapi: 3.0.0
info:
  title: 'TiSaude API'
  version: '0.1'
paths:
  /api/appointments:
    get:
      tags:
        - Consultas
      summary: 'Lista de Consultas'
      description: 'Lista de Consultas'
      operationId: 812c4f15927f3d0e3e959538a6f1104b
      responses:
        '200':
          description: 'Lista de Consultas'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Consultas
      summary: 'Cria um nova Consulta'
      description: 'Cria um nova Consulta'
      operationId: 2b481df28ec53d2afb2ec7c9a3a77846
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Cria um nova Consulta'
      security:
        -
          bearerAuth: []
  '/api/appointments/{id}':
    get:
      tags:
        - Consultas
      summary: 'Exibe uma Consulta'
      description: 'Exibe uma Consulta'
      operationId: cca217dc71eae262de1fe09f314a4948
      parameters:
        -
          name: id
          in: path
          description: 'ID da Consulta'
          required: true
      responses:
        '200':
          description: 'Exibe uma Consulta'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Consultas
      summary: 'Atualiza uma Consulta'
      description: 'Atualiza uma Consulta'
      operationId: 56cd9e193d590ed31b28bf8fe7162e70
      parameters:
        -
          name: name
          in: query
          description: Nome
          required: true
      responses:
        '200':
          description: 'Atualiza uma Consulta'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Consultas
      summary: 'Deleta uma Consulta'
      description: 'Deleta uma Consulta'
      operationId: 47535f80ea6b8f4309694ff2123524d3
      responses:
        '200':
          description: 'Deleta uma Consulta'
      security:
        -
          bearerAuth: []
  '/api/appointments/{id}/add-procedures':
    post:
      tags:
        - Consultas
      summary: 'Adiciona um ou vários procedimentos a uma Consulta'
      description: 'Adiciona um ou vários procedimentos a uma Consulta'
      operationId: 3cbe92b2cdcc26801f4dd51912806250
      parameters:
        -
          name: id
          in: path
          description: 'ID da Consulta'
          required: true
        -
          name: 'procedures_ids[]'
          in: query
          description: 'IDs dos Procedimentos'
          required: true
          schema:
            type: array
            items:
              type: integer
            minItems: 1
      responses:
        '200':
          description: 'Procedimentos adicionadas'
  '/api/doctors/{id}/remove-procedure':
    post:
      tags:
        - Consultas
      summary: 'Remove um procedimento de uma Consulta'
      description: 'Remove um procedimento de uma Consulta'
      operationId: 3fc0decd72a3c83da04272ac38b40530
      parameters:
        -
          name: id
          in: path
          description: 'ID da Consulta'
          required: true
        -
          name: procedure_id
          in: query
          description: 'ID do Procedimento'
          required: true
      responses:
        '200':
          description: 'Procedimento removido'
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: 'Realiza o login na API'
      description: 'Retorna informaçõs do usuário e o TOKEN de acesso'
      operationId: dd51a484b7fa0f97c7680fdbde14cf5f
      parameters:
        -
          name: email
          in: query
          description: 'E-mail de credencial de acesso (user@tisaude.com.br)'
          required: true
        -
          name: password
          in: query
          description: 'Senha de credencial de acesso (tisaude123@)'
          required: true
      responses:
        '200':
          description: 'Retorna informaçõs do usuário e o TOKEN de acesso'
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header